feat: Added Email Notification Service, fix: Changes in userManagement and LibrarianBookManagement for communication between services.

+++ userManagement/src/main/java/com/HashedInLibraryApplication/userManagement/repository/UserManagementRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserManagementRepository extends JpaRepository<User, Integer> {
+public interface UserManagementRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByEmpIdAndRole(Integer empId, String role);
 }
